Affichage scores OK

+++ src/Controller/GameController.java
@@ -2,10 +2,9 @@ package Controller;
 
 import Class.observable.QuoridorGame;
 import Class.*;
-
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
+
 
 public class GameController extends AbstractGameController {
 
@@ -58,7 +57,7 @@ public class GameController extends AbstractGameController {
         return Scores.getTopRank();
     }
 
-    public static LinkedHashMap<String, List<Double>> getTopRankByParticipation() {
+    public static LinkedHashMap<String, List<Integer>> getTopRankByParticipation() {
         if(!Scores.isScores()) return null;
         return Scores.getTopRankByParticipation();
     }
